fix(engine): dtype-preservation + shape-assertion + LUFS tail batch (#3684, #3685, #3686, #3687, #3688)
- simple_mastering crossfade ramps: float64 → input dtype
- ParallelEQProcessor mono FFT output: explicit astype(audio.dtype)
- RealtimeAdaptiveEQ remainder path: shape assertion after backend call
- normalize(): explicit astype to avoid NEP-50 implicit dependency
- AdaptiveLimiter._compute_peak_envelope: np.zeros uses input dtype

+++ auralis/dsp/dynamics/limiter.py
@@ -151,7 +151,11 @@ class AdaptiveLimiter:
         # [i, i + lookahead), giving true lookahead peak detection.
         # The previous negative origin looked *backward*, defeating the
         # purpose of the lookahead delay (mirrors BrickWallLimiter fix #3308).
-        padded = np.concatenate([abs_audio, np.zeros(lookahead)])
+        # #3688: dtype-match the zeros so this path doesn't latently promote
+        # float32 to float64. Currently masked by VectorizedEnvelopeFollower
+        # hard-coding float32 output downstream, but the local promotion is
+        # still wasteful and fragile under envelope-follower substitution.
+        padded = np.concatenate([abs_audio, np.zeros(lookahead, dtype=abs_audio.dtype)])
         peak_envelope = maximum_filter1d(
             padded,
             size=lookahead,
